[AIP-231] Document all AIP-231 rules.
Also changes a rule name in AIP-136 to better be consistent.
