+++ rules/aip0136/http_uri_suffix.go
@@ -25,7 +25,7 @@ import (
 )
 
 var uriSuffix = &lint.MethodRule{
-	Name:   lint.NewRuleName("core", "0136", "uri-suffix"),
+	Name:   lint.NewRuleName("core", "0136", "http-uri-suffix"),
 	OnlyIf: isCustomMethod,
 	LintMethod: func(m *desc.MethodDescriptor) []lint.Problem {
 		for _, httpRule := range utils.GetHTTPRules(m) {
