+++ rules/aip0231/http_uri_suffix.go
@@ -15,24 +15,26 @@
 package aip0231
 
 import (
-	"strings"
-	"testing"
-
 	"github.com/googleapis/api-linter/lint"
-	"github.com/jhump/protoreflect/desc/builder"
+	"github.com/googleapis/api-linter/rules/internal/utils"
+	"github.com/jhump/protoreflect/desc"
 )
 
-func TestAddRules(t *testing.T) {
-	rules := make(lint.RuleRegistry)
-	AddRules(rules)
-	for ruleName := range rules {
-		if !strings.HasPrefix(string(ruleName), "core::0231") {
-			t.Errorf("Rule %s is not namespaced to core::0231.", ruleName)
+// Batch Get methods should have a proper HTTP pattern.
+var uriSuffix = &lint.MethodRule{
+	Name:   lint.NewRuleName("core", "0231", "http-name"),
+	OnlyIf: isBatchGetMethod,
+	LintMethod: func(m *desc.MethodDescriptor) []lint.Problem {
+		// Establish that the RPC has no HTTP body.
+		for _, httpRule := range utils.GetHTTPRules(m) {
+			if !batchGetURINameRegexp.MatchString(httpRule.URI) {
+				return []lint.Problem{{
+					Message:    `Get methods URI should be end with ":batchGet".`,
+					Descriptor: m,
+				}}
+			}
 		}
-	}
-}
 
-type field struct {
-	fieldName string
-	fieldType *builder.FieldType
+		return nil
+	},
 }
