+++ rules/aip0231/request_names_field_test.go
@@ -19,149 +19,7 @@ import (
 
 	"github.com/googleapis/api-linter/rules/internal/testutils"
 	"github.com/jhump/protoreflect/desc"
-	"github.com/jhump/protoreflect/desc/builder"
 )
-
-type field struct {
-	fieldName string
-	fieldType *builder.FieldType
-}
-
-func TestParentField(t *testing.T) {
-	// Set up the testing permutations.
-	tests := []struct {
-		testName     string
-		messageName  string
-		messageField *field
-		problems     testutils.Problems
-		problemDesc  func(m *desc.MessageDescriptor) desc.Descriptor
-	}{
-		{
-			"Valid",
-			"BatchGetBooksRequest",
-			&field{"parent", builder.FieldTypeString()},
-			testutils.Problems{},
-			nil},
-		{
-			"MissingField",
-			"BatchGetBooksRequest",
-			nil,
-			testutils.Problems{{Message: "parent"}},
-			nil,
-		},
-		{
-			"InvalidType",
-			"BatchGetBooksRequest",
-			&field{"parent", builder.FieldTypeDouble()},
-			testutils.Problems{{Message: "string"}},
-			func(m *desc.MessageDescriptor) desc.Descriptor {
-				return m.FindFieldByName("parent")
-			},
-		},
-		{
-			"Irrelevant",
-			"EnumerateBooksRequest",
-			&field{"id", builder.FieldTypeString()},
-			testutils.Problems{},
-			nil,
-		},
-	}
-
-	// Run each test individually.
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			// Create an appropriate message descriptor.
-			messageBuilder := builder.NewMessage(test.messageName)
-
-			if test.messageField != nil {
-				messageBuilder.AddField(
-					builder.NewField(test.messageField.fieldName, test.messageField.fieldType),
-				)
-			}
-
-			message, err := messageBuilder.Build()
-			if err != nil {
-				t.Fatalf("Could not build %s message.", test.messageName)
-			}
-
-			// Determine the descriptor that a failing test will attach to.
-			var problemDesc desc.Descriptor = message
-			if test.problemDesc != nil {
-				problemDesc = test.problemDesc(message)
-			}
-
-			// Run the lint rule, and establish that it returns the correct problems.
-			problems := parentField.Lint(message.GetFile())
-			if diff := test.problems.SetDescriptor(problemDesc).Diff(problems); diff != "" {
-				t.Errorf(diff)
-			}
-		})
-	}
-}
-
-func TestResourceField(t *testing.T) {
-	// Set up the testing permutations.
-	tests := []struct {
-		testName    string
-		src         string
-		problems    testutils.Problems
-		problemDesc func(m *desc.MessageDescriptor) desc.Descriptor
-	}{
-		{
-			testName: "Valid",
-			src: `
-message BatchGetBooksResponse {
-  // Books requested.
-  repeated Book books = 1;
-}`,
-			problems: testutils.Problems{},
-		},
-		{
-			testName: "FieldIsNotRepeated",
-			src: `
-message BatchGetBooksResponse {
-  // Book requested.
-  Book book = 1;
-}`,
-			problems: testutils.Problems{{Message: "The \"Book\" type field on Batch Get Response message should be repeated"}},
-		},
-		{
-			testName: "MissingField",
-			src: `
-message BatchGetBooksResponse {
-  string response = 1;
-}`,
-			problems: testutils.Problems{{Message: "Message \"BatchGetBooksResponse\" has no \"Book\" type field"}},
-			problemDesc: func(m *desc.MessageDescriptor) desc.Descriptor {
-				return m
-			},
-		},
-	}
-
-	// Run each test individually.
-	for _, test := range tests {
-		t.Run(test.testName, func(t *testing.T) {
-			template := `
-{{.Src}}
-message Book {
-}`
-			file := testutils.ParseProto3Tmpl(t, template, struct{ Src string }{test.src})
-
-			m := file.GetMessageTypes()[0]
-
-			// Determine the descriptor that a failing test will attach to.
-			var problemDesc desc.Descriptor = m.GetFields()[0]
-			if test.problemDesc != nil {
-				problemDesc = test.problemDesc(m)
-			}
-
-			problems := resourceField.Lint(file)
-			if diff := test.problems.SetDescriptor(problemDesc).Diff(problems); diff != "" {
-				t.Errorf(diff)
-			}
-		})
-	}
-}
 
 func TestNamesField(t *testing.T) {
 	// Set up the testing permutations.
